requirements.txt
asgiref==3.7.2
Django==4.2.3
django-admin-easy==0.7.0
django-environ==0.10.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
mysqlclient==2.2.0
PyJWT==2.7.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
XlsxWriter==3.1.2
